User authentication
- Payments history

+++ app/Http/Controllers/PaymentsController.php
@@ -2,14 +2,21 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\PaymentForm;
 use App\Payment;
+use App\Http\Requests\PaymentForm;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class PaymentsController extends Controller
 {
+    /**
+     * Display a list of payments.
+     *
+     * @return \Illuminate\View\View
+     */
     public function index()
     {
-        $payments = Payment::whereNull('paid_at')
+        $payments = request()->user()->payments()
+            ->whereNull('paid_at')
             ->orderBy('due_date')
             ->get();
 
@@ -21,6 +28,11 @@ class PaymentsController extends Controller
         ]);
     }
 
+    /**
+     * Display a form to create a payment.
+     *
+     * @return \Illuminate\View\View
+     */
     public function create()
     {
         return view('layout')->with([
@@ -29,6 +41,12 @@ class PaymentsController extends Controller
         ]);
     }
 
+    /**
+     * Store a new payment into database.
+     *
+     * @param  \App\Http\Requests\PaymentForm $request
+     * @return array
+     */
     public function store(PaymentForm $request)
     {
         $payment = $request->store();
@@ -36,6 +54,12 @@ class PaymentsController extends Controller
         return compact('payment');
     }
 
+    /**
+     * Display a form to edit a payment.
+     *
+     * @param  \App\Payment $payment
+     * @return \Illuminate\View\View
+     */
     public function edit(Payment $payment)
     {
         return view('layout')->with([
@@ -47,6 +71,13 @@ class PaymentsController extends Controller
         ]);
     }
 
+    /**
+     * Update a payment.
+     *
+     * @param  \App\Http\Requests\PaymentForm $request
+     * @param  \App\Payment $payment
+     * @return array
+     */
     public function update(PaymentForm $request, Payment $payment)
     {
         $payment = $request->update($payment);
@@ -54,10 +85,20 @@ class PaymentsController extends Controller
         return compact('payment');
     }
 
+    /**
+     * Delete a payment.
+     *
+     * @param  \App\Payment $payment
+     * @return \Illuminate\Contracts\Routing\ResponseFactory|\Symfony\Component\HttpFoundation\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function destroy(Payment $payment)
     {
-        $payment->delete();
+        if (! request()->user()->owns($payment)) {
+            throw new AuthorizationException();
+        }
 
+        $payment->delete();
         return response(['success' => true]);
     }
 }
